Move test to stream package; create abstract test class
